Seperate tile element from map
