--- .vs/CMakeSettings.json
@@ -1,21 +1,0 @@
-{
-    "configurations": [
-        {
-            "name": "x64-Debug",
-            "inheritEnvironments": [ "msvc_x64_x64" ],
-            "generator": "Ninja",
-            "buildRoot": "${env.LOCALAPPDATA}\\CMakeBuild\\${workspaceHash}\\build\\${name}",
-            "configurationType": "Debug",
-            "variables": [
-                {
-                    "name": "VCPKG_TARGET_TRIPLET",
-                    "value": "x64-windows-winssl"
-                },
-                {
-                    "name": "CMAKE_TOOLCHAIN_FILE",
-                    "value": "C:\\vcpkg\\scripts\\buildsystems\\vcpkg.cmake"
-                }
-            ]
-        }
-    ]
-}
