+++ src/openrct2/world/Wall.cpp
@@ -689,12 +689,12 @@ static money32 WallSetColour(sint16 x,
     return 0;
 }
 
-uint8 wall_element_get_animation_frame(const rct_tile_element * wallElement)
+uint8 wall_get_animation_frame(const rct_tile_element * wallElement)
 {
     return (wallElement->properties.wall.animation >> 3) & 0xF;
 }
 
-void wall_element_set_animation_frame(rct_tile_element * wallElement, uint8 frameNum)
+void wall_set_animation_frame(rct_tile_element * wallElement, uint8 frameNum)
 {
     wallElement->properties.wall.animation &= WALL_ANIMATION_FLAG_ALL_FLAGS;
     wallElement->properties.wall.animation |= (frameNum & 0xF) << 3;
