+++ src/openrct2/world/Wall.h
@@ -25,4 +25,6 @@ colour_t wall_get_tertiary_colour(const rct_tile_element * tileElement);
 void wall_set_primary_colour(rct_tile_element * tileElement, colour_t colour);
 void wall_set_secondary_colour(rct_tile_element * wallElement, colour_t secondaryColour);
 void wall_set_tertiary_colour(rct_tile_element * tileElement, colour_t colour);
+uint8 wall_get_animation_frame(const rct_tile_element * fenceElement);
+void wall_set_animation_frame(rct_tile_element * wallElement, uint8 frameNum);
